+++ backend/pullbackup/http_auth.py
@@ -1,3 +1,5 @@
+# SPDX-License-Identifier: AGPL-3.0-or-later
+# Copyright (C) 2026 Mike O'Toole
 import base64
 import binascii
 import hmac